Fix bug when user's language is not in the list
