+++ src/data/languages.js
@@ -1,0 +1,12 @@
+export const languages = {
+    en: { name: "English" },
+    cs: { name: "Čeština" },
+    de: { name: "Deutsche" },
+    es: { name: "Español" },
+    fr: { name: "Français" },
+    ko: { name: "한국어" },
+    pl: { name: "Polski" },
+    ru: { name: "Pусский" },
+    tr: { name: "Türkçe" },
+    zh: { name: "繁體中文" },
+};
